Add option to pop without blocking
Queue#pop's default behavior blocks when the queue is empty, until a
new message comes in.

It optionally takes a boolean non_block argument. When
non_block=true, a ThreadException is thrown when the queue is empty.

This PR allows users to pass non_block=true to the underlying message
Queue.