Relax the re-send messages spec for the producer
The underlying behavior is still tested properly here.
The reason for this is that with new versions of nsqd (v0.3.6) the connection
is taking slightly longer to fail when nsqd goes down. As a result
Socket#write isn't failing until after we think we've sent a handful of
messages. Not really a big deal since we're not promising messages will
get delivered under all circumstances.