Improve implementation of producer, timeout/retry
This commit has been inspired greatly by what is done in the official
golang driver.

To follow the recommandations of the NSQ team, the producer is not able
to get initialized with nsqlookupd URL, a producer is connecting to
only one NSQd instance, not several.

- The producer is keeping a list of transactions to wait for the return
value of NSQd.
- A new class `NsqdsProducer` can be initialized with multiple NSQds
addresses and will apply a strategy to write messages:
  - `Nsq::NsqdsProducer::STRATEGY_FAILOVER` Always send on the same
nsqd except if not available, then get to the next one
  - `Nsq::NsqdsProducer::STRATEGY_ROUND_ROBIN` Apply round robin over
the different available nsqds